+++ packages/firebase-dart/lib/firebase.dart
@@ -4,9 +4,11 @@
 /// See: <https://firebase.google.com/docs/reference/js/firebase>.
 library firebase;
 
+export 'src/analytics.dart';
 export 'src/app.dart';
 export 'src/auth.dart';
 export 'src/database.dart';
 export 'src/messaging.dart';
+export 'src/performance.dart';
 export 'src/storage.dart';
 export 'src/top_level.dart';