+++ packages/firebase-dart/lib/src/interop/firebase_interop.dart
@@ -5,12 +5,14 @@ library firebase.firebase_interop;
 
 import 'package:js/js.dart';
 
+import 'analytics_interop.dart';
 import 'app_interop.dart';
 import 'auth_interop.dart';
 import 'database_interop.dart';
 import 'es6_interop.dart';
 import 'firestore_interop.dart';
 import 'messaging_interop.dart';
+import 'performance_interop.dart';
 import 'storage_interop.dart';
 
 @JS()
@@ -36,6 +38,10 @@ external StorageJsImpl storage([AppJsImpl app]);
 external FirestoreJsImpl firestore([AppJsImpl app]);
 @JS()
 external MessagingJsImpl messaging([AppJsImpl app]);
+@JS()
+external AnalyticsJsImpl analytics([AppJsImpl app]);
+@JS()
+external PerformanceJsImpl performance([AppJsImpl app]);
 
 /// https://firebase.google.com/docs/reference/js/firebase.User
 @anonymous
@@ -130,6 +136,10 @@ class FirebaseOptions {
   external set storageBucket(String s);
   external String get messagingSenderId;
   external set messagingSenderId(String s);
+  external String get measurementId;
+  external set measurementId(String s);
+  external String get appId;
+  external set appId(String s);
 
   external factory FirebaseOptions(
       {String apiKey,
@@ -137,5 +147,7 @@ class FirebaseOptions {
       String databaseURL,
       String projectId,
       String storageBucket,
-      String messagingSenderId});
+      String messagingSenderId,
+      String measurementId,
+      String appId});
 }