+++ packages/firebase-dart/lib/src/top_level.dart
@@ -1,11 +1,15 @@
+import 'dart:html';
+
 import 'package:js/js_util.dart' as js;
 
+import 'analytics.dart';
 import 'app.dart';
 import 'auth.dart';
 import 'database.dart';
 import 'firestore.dart';
 import 'interop/firebase_interop.dart' as firebase;
 import 'messaging.dart';
+import 'performance.dart';
 import 'storage.dart';
 
 export 'interop/firebase_interop.dart' show SDK_VERSION;
@@ -33,7 +37,9 @@ App initializeApp(
     String projectId,
     String storageBucket,
     String messagingSenderId,
-    String name}) {
+    String name,
+    String measurementId,
+    String appId}) {
   name ??= _defaultAppName;
 
   try {
@@ -44,7 +50,9 @@ App initializeApp(
             databaseURL: databaseURL,
             projectId: projectId,
             storageBucket: storageBucket,
-            messagingSenderId: messagingSenderId),
+            messagingSenderId: messagingSenderId,
+            measurementId: measurementId,
+            appId: appId),
         name));
   } catch (e) {
     if (_firebaseNotLoaded(e)) {
@@ -120,6 +128,21 @@ Messaging messaging([App app]) {
       (app != null) ? firebase.messaging(app.jsObject) : firebase.messaging();
 
   return Messaging.getInstance(jsObject);
+}
+
+Analytics analytics([App app]) {
+  var jsObject =
+      (app != null) ? firebase.analytics(app.jsObject) : firebase.analytics();
+
+  return Analytics.getInstance(jsObject);
+}
+
+Performance performance([App app]) {
+  var jsObject = (app != null)
+      ? firebase.performance(app.jsObject)
+      : firebase.performance();
+
+  return Performance.getInstance(jsObject);
 }
 
 /// Exception thrown when the firebase.js is not loaded.